+++ jumi-core/src/main/java/fi/jumi/core/files/TestClassFinderFactory.java
@@ -17,19 +17,24 @@ public class TestClassFinderFactory {
     public static TestClassFinder create(SuiteConfiguration suite, ClassLoader classLoader) {
         List<String> testClasses = suite.testClasses();
         String testFileMatcher = suite.testFileMatcher();
+
         if (!testClasses.isEmpty()) {
+            // TODO: remove enumeration and only use file name patterns?
             return new EnumeratedTestClassFinder(testClasses, classLoader);
+
         } else if (!testFileMatcher.isEmpty()) {
+            List<TestClassFinder> finders = new ArrayList<>();
             for (Path dir : getClassesDirectories(suite)) {
-                return new FileNamePatternTestClassFinder(testFileMatcher, dir, classLoader);
+                finders.add(new FileNamePatternTestClassFinder(testFileMatcher, dir, classLoader));
             }
-            throw new AssertionError("TODO"); // TODO
+            return new CompositeTestClassFinder(finders);
+
         } else {
             throw new IllegalArgumentException("testClasses and testFileMatcher were both empty");
         }
     }
 
-    static List<Path> getClassesDirectories(SuiteConfiguration suite) {
+    private static List<Path> getClassesDirectories(SuiteConfiguration suite) {
         ArrayList<Path> dirs = new ArrayList<>();
         for (URI uri : suite.classPath()) {
             Path path = Paths.get(uri);